Added the option to specify the platforms where the smart banner needs to be shown
Sometimes, you only want to show the Smart banner on specific devices and not show them on others e.g. I want to show the banner on iOS and Android but not on Windows or Kindle platforms. I can specify that as an option and then I won't see the Smart banner on other supported platforms.

Example:

$.smartbanner({
        platforms: ['ios', 'android']
});